RE-033: Organize API into classes, bind API origin to .env
