+++ server/.env
@@ -1,1 +1,3 @@
 PORT=8000
+API_ORIGIN=http://109.120.190.243:8080
+API_VERSION=v1
