+++ static/src/pages/sign-in.js
@@ -4,7 +4,7 @@ import { Component } from "../component.js";
 import { loadPath } from "../main.js";
 import { InputField } from "../components/input-field.js";
 import { Button } from "../components/button.js";
-import { API } from "../api.js";
+import { API } from "../modules/api.js";
 
 export class SignInPage extends Component {
     get pageRoot() {
