+++ static/src/pages/sign-up.js
@@ -1,7 +1,7 @@
 'use strict';
 
 import { Component } from "../component.js";
-import { API } from "../api.js";
+import { API } from "../modules/api.js";
 import { loadPath } from "../main.js";
 import { InputField } from "../components/input-field.js";
 import { Button } from "../components/button.js";
